distutils.bbclass: only remove if existing

+++ classes/distutils.bbclass
@@ -51,7 +51,7 @@ distutils_do_install() {
             done
         fi
 
-        if test -e ${D}${sbindir} ; then
+        if test -e ${D}${sbindir}; then
             for i in ${D}${sbindir}/* ; do \
                 sed -i -e s:${STAGING_BINDIR_NATIVE}:${bindir}:g $i
             done
@@ -62,7 +62,9 @@ distutils_do_install() {
         #
         # FIXME: Bandaid against wrong datadir computation
         #
-        mv -f ${D}${datadir}/share/* ${D}${datadir}/
+        if test -e ${D}${datadir}/share; then
+            mv -f ${D}${datadir}/share/* ${D}${datadir}/
+        fi
 }
 
 EXPORT_FUNCTIONS do_compile do_install
